Move utils tests to own folder
